TDI-30195 : refactor the AbstractPublishJobAction

+++ main/plugins/org.talend.designer.publish.core/src/org/talend/designer/publish/core/actions/AbstractPublishJobAction.java
@@ -81,52 +81,40 @@ public abstract class AbstractPublishJobAction implements IRunnableWithProgress 
 
     @Override
     public final void run(IProgressMonitor monitor) throws InvocationTargetException, InterruptedException {
-        File tmpJob;
-        try {
-            switch (exportType) {
-            case POJO:
-                tmpJob = File.createTempFile("item", ".zip", null);
-                break;
-            case OSGI:
-                tmpJob = File.createTempFile("job", ".jar", null);
-                break;
-            default:
-                tmpJob = File.createTempFile("job", ".jar", null);
-                break;
-            }
-        } catch (IOException e) {
-            throw new InvocationTargetException(e);
+        switch (exportType) {
+        case POJO:
+            exportJobForPOJO(monitor);
+            break;
+        case OSGI:
+            exportJobForOSGI(monitor);
+            break;
+        default:
+            exportJobForOSGI(monitor);
+            break;
         }
-        JobExportAction action = null;
-        if (exportType == JobExportType.OSGI) {
+        // http://jira.talendforge.org/browse/TESB-5426 LiXiaopeng
+        monitor.done();
+    }
+
+    private void exportJobForOSGI(IProgressMonitor monitor) throws InvocationTargetException, InterruptedException {
+        File tmpJob = null;
+        try {
+            tmpJob = File.createTempFile("job", ".jar", null);
             jobScriptsManager = JobScriptsManagerFactory.createManagerInstance(
                     JobScriptsManagerFactory.getDefaultExportChoiceMap(), IContext.DEFAULT, JobScriptsManager.LAUNCHER_ALL,
                     IProcessor.NO_STATISTICS, IProcessor.NO_TRACES, JobExportType.OSGI);
             // generate
             jobScriptsManager.setDestinationPath(tmpJob.getAbsolutePath());
-            action = new JobExportAction(Collections.singletonList(node), jobVersion, bundleVersion, jobScriptsManager,
-                    System.getProperty("java.io.tmpdir"));
-        } else {
-            jobScriptsManager.setDestinationPath(tmpJob.getAbsolutePath());
-            action = new JobExportAction(Collections.singletonList(node), jobVersion, jobScriptsManager, null, "job");
-        }
+            JobExportAction action = new JobExportAction(Collections.singletonList(node), jobVersion, bundleVersion,
+                    jobScriptsManager, System.getProperty("java.io.tmpdir"));
 
-        action.run(monitor);
-        if (!action.isBuildSuccessful()) {
-            return;
-        }
-
-        // publish
-        monitor.beginTask("Deploy to Artifact Repository....", IProgressMonitor.UNKNOWN);
-        // http://jira.talendforge.org/browse/TESB-5426 LiXiaopeng
-        try {
+            action.run(monitor);
+            if (!action.isBuildSuccessful()) {
+                return;
+            }
+            monitor.beginTask("Deploy to Artifact Repository....", IProgressMonitor.UNKNOWN);
             ProcessItem processItem = (ProcessItem) node.getObject().getProperty().getItem();
-
-            FeaturesModel featuresModel = new FeaturesModel(groupId, artifactName, artifactVersion);
-            featuresModel.setConfigName(node.getObject().getLabel());
-            featuresModel.setContexts(JobContextUtils.getContextsMap(processItem));
-            BundleModel bundleModel = new BundleModel(groupId, artifactName, artifactVersion, tmpJob);
-            featuresModel.addBundle(bundleModel);
+            FeaturesModel featuresModel = getFeatureModel(tmpJob);
 
             // [TESB-12036] add talend-data-mapper feature
             NodeType tHMapNode = EmfModelUtils.getComponentByName(processItem, THMAP_COMPONENT_NAME);
@@ -134,22 +122,53 @@ public abstract class AbstractPublishJobAction implements IRunnableWithProgress 
                 featuresModel.addFeature(new FeatureModel(FeaturesModel.TALEND_DATA_MAPPER_FEATURE_NAME,
                         FeaturesModel.ESB_FEATURE_VERSION_RANGE));
             }
-
             Collection<NodeType> tIPaasComponents = EmfModelUtils.getComponentsByName(processItem, "tActionInput",
                     "tActionOutput");
             if (!tIPaasComponents.isEmpty() && exportType == JobExportType.OSGI) {
                 addMissingBundles(featuresModel, ((JobJavaScriptOSGIForESBManager) jobScriptsManager).getExcludedModuleNeededs());
             }
+            process(processItem, getFeatureModel(tmpJob), monitor);
+        } catch (IOException e) {
+            throw new InvocationTargetException(e);
+        } finally {
+            if (tmpJob != null && tmpJob.exists()) {
+                tmpJob.delete();
+            }
+        }
+    }
 
+    private void exportJobForPOJO(IProgressMonitor monitor) throws InvocationTargetException, InterruptedException {
+        File tmpJob = null;
+        try {
+            tmpJob = File.createTempFile("item", ".zip", null);
+            jobScriptsManager.setDestinationPath(tmpJob.getAbsolutePath());
+            JobExportAction action = new JobExportAction(Collections.singletonList(node), jobVersion, jobScriptsManager, null,
+                    "job");
+            action.run(monitor);
+            if (!action.isBuildSuccessful()) {
+                return;
+            }
+            monitor.beginTask("Deploy to Artifact Repository....", IProgressMonitor.UNKNOWN);
+            ProcessItem processItem = (ProcessItem) node.getObject().getProperty().getItem();
+            FeaturesModel featuresModel = getFeatureModel(tmpJob);
             process(processItem, featuresModel, monitor);
         } catch (IOException e) {
             throw new InvocationTargetException(e);
         } finally {
-            if (tmpJob.exists()) {
+            if (tmpJob != null && tmpJob.exists()) {
                 tmpJob.delete();
             }
         }
-        monitor.done();
+    }
+
+    private FeaturesModel getFeatureModel(File tmpJob) {
+        ProcessItem processItem = (ProcessItem) node.getObject().getProperty().getItem();
+        FeaturesModel featuresModel = new FeaturesModel(groupId, artifactName, artifactVersion);
+        featuresModel.setConfigName(node.getObject().getLabel());
+        featuresModel.setContexts(JobContextUtils.getContextsMap(processItem));
+        BundleModel bundleModel = new BundleModel(groupId, artifactName, artifactVersion, tmpJob);
+        featuresModel.addBundle(bundleModel);
+        return featuresModel;
     }
 
     protected String getJobVersion() {
